Moved PassKeyVerifier
